+++ packages/@expo/cli/src/start/server/metro/withMetroMultiPlatform.ts
@@ -143,8 +143,9 @@ export function withExtendedResolver(
   }
 ) {
   if (isFastResolverEnabled) {
-    Log.warn(`Experimental bundling features are enabled.`);
+    Log.warn(`Experimental module resolution is enabled.`);
   }
+
   if (isReactCanaryEnabled) {
     Log.warn(`Experimental React Canary version is enabled.`);
   }