+++ packages/@expo/cli/src/start/server/middleware/metroOptions.ts
@@ -33,6 +33,8 @@ export type ExpoMetroOptions = {
   inlineSourceMap?: boolean;
   clientBoundaries?: string[];
   splitChunks?: boolean;
+  /** Enable optimized bundling (tree shaking). */
+  optimize?: boolean;
 };
 
 export type SerializerOptions = {
@@ -77,10 +79,17 @@ function withDefaults({
     }
   }
 
+  const optimize =
+    props.optimize ??
+    (props.environment !== 'node' &&
+      mode === 'production' &&
+      env.EXPO_UNSTABLE_METRO_OPTIMIZE_GRAPH);
+
   return {
     mode,
     minify,
     preserveEnvVars,
+    optimize,
     lazy: !props.isExporting && lazy,
     ...props,
   };
@@ -141,6 +150,7 @@ export function getMetroDirectBundleOptions(
     inlineSourceMap,
     splitChunks,
     reactCompiler,
+    optimize,
   } = withDefaults(options);
 
   const dev = mode !== 'production';
@@ -175,6 +185,7 @@ export function getMetroDirectBundleOptions(
     unstable_transformProfile: isHermes ? 'hermes-stable' : 'default',
     customTransformOptions: {
       __proto__: null,
+      optimize: optimize || undefined,
       engine,
       preserveEnvVars: preserveEnvVars || undefined,
       // Use string to match the query param behavior.
@@ -240,6 +251,7 @@ export function createBundleUrlSearchParams(options: ExpoMetroOptions): URLSearc
     isExporting,
     clientBoundaries,
     splitChunks,
+    optimize,
   } = withDefaults(options);
 
   const dev = String(mode !== 'production');
@@ -302,6 +314,9 @@ export function createBundleUrlSearchParams(options: ExpoMetroOptions): URLSearc
 
   if (splitChunks) {
     queryParams.append('serializer.splitChunks', String(splitChunks));
+  }
+  if (optimize) {
+    queryParams.append('transform.optimize', String(optimize));
   }
   if (serializerOutput) {
     queryParams.append('serializer.output', serializerOutput);