+++ packages/@expo/metro-runtime/src/error-overlay/modules/ExceptionsManager/index.native.ts
@@ -1,2 +1,3 @@
 import ExceptionsManager from 'react-native/Libraries/Core/ExceptionsManager';
+
 export default ExceptionsManager;