+++ packages/expo-modules-core/android/src/main/cpp/types/JNIToJSIConverter.cpp
@@ -3,6 +3,8 @@
 #include "JNIToJSIConverter.h"
 #include "../JavaReferencesCache.h"
 #include "ObjectDeallocator.h"
+#include "../JSharedObject.h"
+#include "../JNIUtils.h"
 
 #include <react/jni/ReadableNativeMap.h>
 #include <react/jni/ReadableNativeArray.h>
@@ -128,15 +130,9 @@ jsi::Value convert(
     cache->getJClass(
       "expo/modules/kotlin/sharedobjects/SharedObject").clazz
   )) {
-    auto jsObject = std::make_shared<jsi::Object>(jsi::Object(rt));
     JSIContext *jsiContext = getJSIContext(rt);
-    auto jsObjectRef = JavaScriptObject::newInstance(
-      jsiContext,
-      jsiContext->runtimeHolder,
-      jsObject
-    );
-    jsiContext->registerSharedObject(jni::make_local(unpackedValue), jsObjectRef);
-    return jsi::Value(rt, *jsObject);
+    auto sharedObject = jni::static_ref_cast<JSharedObject::javaobject>(value);
+    return convertSharedObject(sharedObject, rt, jsiContext);
   }
   if (env->IsInstanceOf(
     unpackedValue,