+++ packages/expo-router/src/fork/getPathFromState.ts
@@ -267,7 +267,7 @@ function walkConfigItems(
     pattern = inputPattern;
 
     if (route.params) {
-      if (route.params['#']) {
+      if (route.params['#'] !== undefined) {
         hash = route.params['#'];
         delete route.params['#'];
       }